dbtest handles all database test prep
The dbtest package is intended to wrap all test database behaviour.
Before this existed, the migrations package had accrued some test
helpers that should no longer live there.

Bring that behaviour into the dbtest package.
