+++ pkg/subscription/integration/suite_test.go
@@ -5,9 +5,6 @@ import (
 
 	kitlog "github.com/go-kit/kit/log"
 
-	// Ensure we migrate the database before running any tests
-	_ "github.com/lawrencejones/pgsink/pkg/migration"
-
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
